refine code for hash join exec
